requirements.txt
Faker==30.3.0
python-dotenv
